Send header and receive response
Current attempts by the client to close a socket for writing, as a
trigger for the server to send a response, fail because ssl:shutdown/2
apparently doesn't really support such a feature, in spite of the
documentation.

This allows the sender to prepend the file with a header containing the
expected file size, which the receiver can then use to determine when it
should stop reading from the socket, and so trigger a response. The
receiver determines whether the stream contains a header by examining
the first 4 bytes for a "magic" byte sequence. Currently, only the
'filesize' key is required, whose value must be a positive integer
matching the byte count of the file.

The response may contain potentially useful information such as the
absolute path of the remote file, received file size, and checksum.